add field_link_resolver to single record view
