+++ app/controllers/registers_controller.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
 class RegistersController < ApplicationController
-  include ActionView::Helpers::UrlHelper
-  helper_method :field_link_resolver
-
   def index
     @registers = Register.available
                          .in_beta
@@ -24,25 +21,6 @@ class RegistersController < ApplicationController
     @register = Register.find_by_slug!(params[:id])
     @records = recover_records(@register.fields_array, params)
     @feedback = Feedback.new
-  end
-
-  def field_link_resolver(field, field_value, register_slug = @register.slug)
-    single_resolver = lambda { |f, fv|
-      if f['datatype'] == 'curie' && fv.include?(':')
-        curie = fv.split(':')
-        curie[1].present? ? link_to(fv, register_record_path(curie[0], curie[1])) : link_to(fv, register_path(register_slug))
-      elsif f['register'].present? && f['field'] != register_slug
-        link_to(fv, register_record_path(f['register'], fv))
-      elsif f['datatype'] == 'url'
-        link_to(fv, fv)
-      else
-        fv
-      end
-    }
-
-    cardinality_n_links = -> { field_value.map { |fv| single_resolver.call(field, fv) }.join(', ').html_safe }
-
-    field_value.is_a?(Array) ? cardinality_n_links.call : single_resolver.call(field, field_value)
   end
 
 private
