+++ app/helpers/application_helper.rb
@@ -1,4 +1,6 @@
 module ApplicationHelper
+  include ActionView::Helpers::UrlHelper
+  
   def crest_class_name(authority)
     case authority
     when "home-office"
@@ -53,4 +55,24 @@ module ApplicationHelper
                                                  class: css_class
     end
   end
+
+  def field_link_resolver(field, field_value, register_slug = @register.slug)
+    single_resolver = lambda { |f, fv|
+      if f['datatype'] == 'curie' && fv.include?(':')
+        curie = fv.split(':')
+        curie[1].present? ? link_to(fv, register_record_path(curie[0], curie[1])) : link_to(fv, register_path(register_slug))
+      elsif f['register'].present? && f['field'] != register_slug
+        link_to(fv, register_record_path(f['register'], fv))
+      elsif f['datatype'] == 'url'
+        link_to(fv, fv)
+      else
+        fv
+      end
+    }
+
+    cardinality_n_links = -> { field_value.map { |fv| single_resolver.call(field, fv) }.join(', ').html_safe }
+
+    field_value.is_a?(Array) ? cardinality_n_links.call : single_resolver.call(field, field_value)
+  end
+
 end
