+++ app/views/records/show.html.haml
@@ -22,7 +22,7 @@
               %td
                 = field['field']
               %td
-                =  field_value
+                =  field_link_resolver(field, field_value)
   .grid-row
     .column-two-thirds
       %p This is a record from the #{link_to @register.name, register_path(@register.slug)} register
